Move RALUpdater to infra module
